modules/test-app/requirements.txt
Flask>=2.2
boto3>=1.26.0
psycopg2-binary>=2.9.5
